[WIP] Split push to applications in logic and gui (#4110)
* Split push to applications in logic and gui

* Properly implement settings factory
