+++ src/main/java/org/jabref/gui/push/PushToEmacs.java
@@ -26,7 +26,6 @@ import org.slf4j.LoggerFactory;
 public class PushToEmacs extends AbstractPushToApplication implements PushToApplication {
 
     private static final Logger LOGGER = LoggerFactory.getLogger(PushToEmacs.class);
-    private final JTextField additionalParams = new JTextField(30);
 
     public PushToEmacs(DialogService dialogService) {
         super(dialogService);
@@ -40,27 +39,6 @@ public class PushToEmacs extends AbstractPushToApplication implements PushToAppl
     @Override
     public JabRefIcon getIcon() {
         return IconTheme.JabRefIcons.APPLICATION_EMACS;
-    }
-
-    @Override
-    public JPanel getSettingsPanel() {
-        additionalParams.setText(Globals.prefs.get(JabRefPreferences.EMACS_ADDITIONAL_PARAMETERS));
-        return super.getSettingsPanel();
-    }
-
-    @Override
-    public void storeSettings() {
-        super.storeSettings();
-        Globals.prefs.put(JabRefPreferences.EMACS_ADDITIONAL_PARAMETERS, additionalParams.getText());
-    }
-
-    @Override
-    protected void initSettingsPanel() {
-        super.initSettingsPanel();
-        builder.appendRows("2dlu, p, 2dlu, p");
-        builder.add(Localization.lang("Additional parameters") + ":").xy(1, 3);
-        builder.add(additionalParams).xy(3, 3);
-        settings = builder.build();
     }
 
     @Override
